+++ src/unit/triple_mesh_test.cc
@@ -19,8 +19,8 @@
 using namespace std;
 using namespace mesh;
 
-static constexpr size_t StrLen = 128;
-static constexpr size_t ObjCount = 32;
+static constexpr uint32_t StrLen = 128;
+static constexpr uint32_t ObjCount = 32;
 
 static char *s1;
 static char *s2;
